add: Utils namespace instead of static class
