+++ inc/ClientInfo.hpp
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ClientInfo.hpp                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sgiochal <sgiochal@student.42.fr>          +#+  +:+       +#+        */
+/*   By: abied-ch <abied-ch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/10 10:35:13 by sgiochal          #+#    #+#             */
-/*   Updated: 2024/03/11 07:30:52 by sgiochal         ###   ########.fr       */
+/*   Updated: 2024/03/12 20:45:55 by abied-ch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef CLIENTINFO_HPP
 #define CLIENTINFO_HPP
+#pragma once
 
-#include <iostream>
-#include "Node.hpp"
+#include "Headers.hpp"
 
 class ClientInfo : public Node 
 {
