--- inc/ConfigUtils.hpp
@@ -1,14 +1,0 @@
-#ifndef CONFIGUTILS_HPP
-#define CONFIGUTILS_HPP
-#pragma once
-
-#include "Headers.hpp"
-
-class ConfigUtils
-{
-	public:
-
-		static bool		file_exists(const std::string& str);
-};
-
-#endif
