+++ inc/Log.hpp
@@ -3,23 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   Log.hpp                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sgiochal <sgiochal@student.42.fr>          +#+  +:+       +#+        */
+/*   By: abied-ch <abied-ch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/11 19:25:56 by sgiochal          #+#    #+#             */
-/*   Updated: 2024/03/12 18:08:43 by sgiochal         ###   ########.fr       */
+/*   Updated: 2024/03/12 20:50:57 by abied-ch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LOG_HPP 
 #define LOG_HPP
+#pragma once
 
-#include <iostream>
-#include <string>
-#include <sstream>
-#include <map>
-#include <ctime>
-#include <fstream>
-
+#include "Headers.hpp"
 
 enum LogDest
 {
